Made more testing of collect and calculate, all tests passed

+++ serverside/server/src/main/java/no/auke/mg/channel/models/ChannelStatus.java
@@ -6,7 +6,7 @@ public class ChannelStatus extends PersistObject {
 	private long starttime;
 	private long currenttime = System.currentTimeMillis();
 	private long hits=0;
-	private int currentpos=0;
+	private long currentpos=0;
 	private int slotTime;
 	private int timeframes;
 	private int usersessions;
@@ -15,14 +15,14 @@ public class ChannelStatus extends PersistObject {
 	public long getStarttime() {return starttime;}
 	public long getCurrenttime() {return currenttime;}
 	public long getHits() {return hits;}
-	public int getCurrentpos() {return currentpos;}
+	public long getCurrentpos() {return currentpos;}
 	public int getSlotTime() {return slotTime;}
 	public int getTimeframes() {return timeframes;}
 	public int getUsersessions() {return usersessions;}
 
 	public void setChannelid(String channelid) {this.channelid = channelid;}
 	public void setStarttime(long starttime) {this.starttime = starttime;}
-	public void setCurrentpos(int currentpos) {this.currentpos = currentpos;}
+	public void setCurrentpos(long currentpos) {this.currentpos = currentpos;}
 	public void setSlotTime(int slotTime) {this.slotTime = slotTime;}
 	public void setTimeframes(int timeframes) {this.timeframes = timeframes;}
 	public void setUsersessions(int usersessions) {this.usersessions = usersessions;}
@@ -34,12 +34,8 @@ public class ChannelStatus extends PersistObject {
 	public void setEventid(String eventid) {this.eventid = eventid;}
 
 	@Override
-	public String getPersistName() {
-		return getChannelid();
-	}
+	public String getPersistName() {return getChannelid();}
 	@Override
-	public String getPersistDir() {
-		return "";
-	}
+	public String getPersistDir() {return "";}
 
 }
